+++ packages/yumemi_lints/CHANGELOG.md
@@ -36,6 +36,7 @@ Examples of version updates are as follows:
 - Changed to be independent of the `pub_semver` package.
 - Changed to be independent of the `yaml` package.
 - Changed to be independent of the `file` package.
+- Changed to be independent of the `args` package.
 
 ## 2.2.0